@@ -13,7 +13,7 @@
 
 	<properties>
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-		<java.version>1.8</java.version>
+		<java.version>11</java.version>
 		<!-- 依赖版本 -->
 		<maven-jar-plugin.version>3.1.1</maven-jar-plugin.version><!--解决pom在springboot高版本报unknown 
 			error -->
@@ -64,7 +64,7 @@
 		    <version>1.4.0</version>
 		</dependency>
 
-		<!-- excel poi -->
+		<!-- excel poi >
 		<dependency>
 			<groupId>org.apache.poi</groupId>
 			<artifactId>poi</artifactId>
@@ -74,7 +74,21 @@
 			<groupId>org.apache.poi</groupId>
 			<artifactId>poi-ooxml</artifactId>
 			<version>3.13</version>
+		</dependency-->
+
+		<!-- https://mvnrepository.com/artifact/org.apache.poi/poi-ooxml -->
+		<dependency>
+		    <groupId>org.apache.poi</groupId>
+		    <artifactId>poi</artifactId>
+		    <version>4.1.2</version>
 		</dependency>
+		<dependency>
+		    <groupId>org.apache.poi</groupId>
+		    <artifactId>poi-ooxml</artifactId>
+		    <version>4.1.2</version>
+		</dependency>
+
+		
 
 		<!-- mybatis -->
 		<dependency>
@@ -115,13 +129,6 @@
 		    <groupId>org.apache.httpcomponents</groupId>
 		    <artifactId>httpclient</artifactId>
 		    <version>4.5.13</version>
-		</dependency>
-		
-		<!-- https://mvnrepository.com/artifact/org.apache.shiro/shiro-spring -->
-		<dependency>
-		    <groupId>org.apache.shiro</groupId>
-		    <artifactId>shiro-spring</artifactId>
-		    <version>1.4.0</version>
 		</dependency>
 
 		<!-- swagger -->
@@ -173,7 +180,7 @@
 
 
 	<build>
-		<finalName>kph</finalName>
+		<finalName>ep</finalName>
 		<resources>
 			<resource>
 				<filtering>true</filtering>
