@@ -25,9 +25,9 @@ export default ()=>{
             3: {
                 text: '医保',
             },
-            // 4: {
-            //     text: '省医保',
-            // },
+            4: {
+                text: '异地医保',
+            },
             5: {
                 text: '现金',
             }
