@@ -6,7 +6,7 @@ import javax.validation.constraints.Pattern;
 public class LoginBo {
 
 	@NotBlank
-	@Pattern(regexp="^[1](([3|5|8][\\d])|([4][4,5,6,7,8,9])|([6][2,5,6,7])|([7][^9])|([9][1,8,9]))[\\d]{8}$")
+	@Pattern(regexp="^[1](([3|5|8][\\d])|([4][4,5,6,7,8,9])|([6][2,5,6,7])|([7][^9])|([9][1,8,9]))[\\d]{8}$", message="请输入正确的手机号")
 	private String phone;
 	
 	@NotBlank
