@@ -37,7 +37,7 @@ public class Bill {
 	public static final int WXPAY = 1;
 	public static final int ZFB = 2;
 	public static final int SHIYIBAO = 3;
-	public static final int SHENGYIBAO = 4;
+	public static final int YIDIYIBAO = 4;
 	public static final int CASH = 5; 
 	
 	@Column(name = "payway")
